Create class Settings to manage game settings
There is no easy way to modify the game's settings.

Each setting must be manually modified and updated.

This makes maintaing the game code difficult and repetitive.

Create and implement the Settings class.

Update the stars.py program to utilize the Settings class.

The Settings class provides a centralized object to manage all settings.

The settings of the game are now easily maintained.